feat: add ingredients to a recipe

+++ TrackCaloriesBot/src/Entity/Recipe.cs
@@ -19,7 +19,7 @@ public class Recipe
     public double BaseProtein { get; set; }
     public double BaseFat { get; set; }
     public double BaseCarbs { get; set; }
-    public int WeightPerServing { get; set; }
+    public double WeightPerServing { get; set; }
     public ICollection<Product>? Products { get; set; }
     public string? Description { get; set; }
 }
